change DMatrix template from i16 to f64

+++ benches/convolve.rs
@@ -8,7 +8,7 @@ use mercer_research::utils::kernel::{
 };
 use nalgebra::DMatrix;
 
-fn get_image_data() -> DMatrix<i16> {
+fn get_image_data() -> DMatrix<f64> {
     get_pixel_matrix(
         &ImageReader::open("images\\mnist_png\\train\\4\\2.png")
             .unwrap()
@@ -20,35 +20,35 @@ fn get_image_data() -> DMatrix<i16> {
 
 pub fn simple_convolution(c: &mut Criterion) {
     c.bench_function("Simple Convolution", |b| {
-        let img_data: DMatrix<i16> = get_image_data();
+        let img_data: DMatrix<f64> = get_image_data();
         b.iter(|| img_data.convolve_2d(&__TOP_SOBEL, &Padding::None));
     });
 }
 
 pub fn separated_convolution(c: &mut Criterion) {
     c.bench_function("Separated Convolution", |b| {
-        let img_data: DMatrix<i16> = get_image_data();
+        let img_data: DMatrix<f64> = get_image_data();
         b.iter(|| img_data.convolve_2d_separated(SeparableOperator::Top, &Padding::None));
     });
 }
 
 pub fn padded_simple(c: &mut Criterion) {
     c.bench_function("Simple Convolution (Padded: Same)", |b| {
-        let img_data: DMatrix<i16> = get_image_data();
+        let img_data: DMatrix<f64> = get_image_data();
         b.iter(|| img_data.convolve_2d(&__TOP_SOBEL, &Padding::Same));
     });
 }
 
 pub fn padded_separated(c: &mut Criterion) {
     c.bench_function("Separated Convolution (Padded: Same)", |b| {
-        let img_data: DMatrix<i16> = get_image_data();
+        let img_data: DMatrix<f64> = get_image_data();
         b.iter(|| img_data.convolve_2d_separated(SeparableOperator::Top, &Padding::Same));
     });
 }
 
 pub fn max_pooling(c: &mut Criterion) {
     c.bench_function("Max Pooling", |b| {
-        let img_data: DMatrix<i16> = get_image_data();
+        let img_data: DMatrix<f64> = get_image_data();
         b.iter(|| img_data.pool_2d(&Padding::Same, &Pooling::Max));
     });
 }
